little changes on model and server configs

+++ db/migrations/20230621141057_create_travel_plans.cr
@@ -2,7 +2,7 @@ class CreateTravelPlans::V20230621141057 < Avram::Migrator::Migration::V1
   def migrate
     # Learn about migrations at: https://luckyframework.org/guides/database/migrations
     create table_for(TravelPlan) do
-      primary_key id : Int32
+      primary_key id : Int64
       add travel_stops : JSON::Any
       add_timestamps
     end
